refactored custom analysis API
